light version footer added
