+++ src/app/page.tsx
@@ -1,27 +1,23 @@
 // import Login from "@/src/components/login/Login";
-import Footer from "../components/footer_test/Footer";
+import Footer from "../components/footer/Footer";
 
 export default function Home() {
   const rateMyStudySpot = {
-    instagram: "https://www.instagram.com/your_instagram",
-    discord: "https://discord.gg/your_discord",
-    github: "https://github.com/your_github",
+    instagram: "",
+    discord: "",
+    github: "",
     logo: "/logo192.png",
   };
   return (
     <main className="relative flex min-h-screen items-center justify-center overflow-x-hidden">
+      {/* Light footer -Dric */}
       <Footer
-        RateMyStudySpot={{
-          instagram: "https://instagram.com/yourprofile",
-          discord: "https://discord.gg/yourinvite",
-          github: "https://github.com/yourprofile",
-          logo: "/logo192.png",
-        }}
-        bgColor="bg-white"
-        textColor="text-black" // Fresh green text
-        textHoverColor="hover:text-blue-600" // More intense green on hover
-        date_text_colour="text-black"
-        company_text_colour="text-blue-400"
+        RateMyStudySpot={rateMyStudySpot}
+        bgFooter="bg-blue-50 border border-blue-200 rounded-md p-4"
+        footerLink="text-gray-900"
+        onLinkHover="transition-all duration-300 ease-in-out hover:text-sky-600 hover:scale-105"
+        dateText="text-black"
+        logoText="relative text-transparent bg-clip-text bg-black"
       />
     </main>
   );
